+++ bot/utils.py
@@ -260,42 +260,3 @@ class DelayedExpressionEvaluator:
 		for (key, value) in self.global_vars.items():
 			locals[key] = value
 		return locals
-
-def filterParametersWithCase(
-	params_from_func: Dict[str, object]
-) -> Dict[str, Case]:
-	params_with_case: Dict[str, Case] = {}
-	for (param, object) in params_from_func.items():
-		if isinstance(object, Case):
-			params_with_case[param] = object
-	return params_with_case
-
-def filterFixtures(
-	params_and_fixtures: Dict[str, Any],
-	params_with_case: Dict[str, Case]
-) -> Dict[str, Any]:
-	fixtures: Dict[str, Case] = params_and_fixtures.copy()
-	for param_key in params_with_case.keys():
-		fixtures.pop(param_key)
-	return fixtures
-
-def evalAndWriteDelayedExprInParams(
-	cases: List[Case],
-	fixtures: Dict[str, Any]
-) -> None:
-	"""
-	The func acts on Case directly via a reference.
-	Args:
-		cases(List[Case]): any `Case` instances that located in Dict with params
-		as value (e.x `pytest.Function.callspec.params <https://docs.pytest.org/
-		en/7.1.x/reference/reference.html#pytest.Function>`).
-	"""
-	for case in cases:
-		params_with_delayed_expr = case.getDelayedExprs()
-		params_with_undelayed_expr: Dict[str, Any] = {}
-		for (param, its_delay_exp) in params_with_delayed_expr.items():
-			eval_results = DelayedExpressionEvaluator(
-				[its_delay_exp],
-				fixtures).eval()
-			params_with_undelayed_expr[param] = eval_results
-		case.setUndelayedExprs(params_with_undelayed_expr)
